pyproject.toml
[project]
name = "micro-payment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crcmod>=1.7",
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "qrcode[pil]>=8.2",
    "redis>=7.1.0",
]
